[Refactor] Comment 도메인 Converter 분리
